Keep compare_or_generate_golden_matrix in test files; reuse repr_value
Move compare_or_generate_golden_matrix back to test files instead of the
shared mixin. Add repr_pandas_series_value to GoldenFileTestMixin so
repr_value handles pd.Series, removing the need for a custom _repr_result.
